Another significant change to layout
I moved alt/meta down on the other side of super/gui/win.

I made the bucky bits 'momentary' again rather than oneshots.  The
downside to oneshots is the 'lock' mode, which works great with layer
switchers but not with bucky bits.  My keyboard would get locked into
super or ctrl mode with no indication on how to get out.

A fix would be to set layer lock for OSLs at the current 2, but be able
to set it to like 9999 for OSMs.