Major reorganization of layers
"CURSR" is about right, that's what I was doing...

(Hmm, I've already seen a small problem when typing 'ut' fast, I got
F11...)

I'm mostly happy with the placement of letter, number, and symbol keys.
I'm not happy with layer and modifier keys :P  Even this layout is not
ideal.

I've discovered that my hands are way too small for an Ergodox.  The
four outermost 1u keys in the thumb cluster are useless to me, as are
the 1u keys next to 5 and 6.  Same story with the three keys below the
Shift-Z-X row, and the equivalents on the right hand.  That's 16 keys
that are hard or painful to use.  So it's clear I may need to
investigate chording (or getting gud at using the bottommost keys).

I could sell this self-made Ergodox, but probably not come close to
recouping my investment.  To someone else, it's a custom product, a
silly objet d'art.  One of its LEDs doesn't work, because I'm not good
at soldering.  No one would find value in it, like they would a packaged
Ergodox EZ.  To me, it was an attempt to make something that I could fix
myself.  It was important to self-build because I wanted to be able to
replace it should something break, so it's probably worth it for me to
hang on to this keyboard and try to figure out how to use it.

(I'd be interested in trying to switch keyboards if I got the ability to
make my own PCBs and got my 3d printer working or a CNC.  None of this
is in my near future.)

The best route, as far as I'm concerned, is going to use this Ergodox
like it's an Atreus or similar --- with minimal keys.  This is a start
in that vein.

---

The change here fixes an issue I was having, where I was routinely
having problems holding down space, backspace, or enter (to get many
newlines, spaces; or to delete many things) and getting nothing.  So
I've moved my layer shifts onto home row keys.

I split the layer shifts so that function keys and administrative
stuff like Teensy reset live on one layer, and cursor navigation
on another.

I use the cursor stuff the most, that lives on 'F'.

---

Future plans:

 - I might move the modifiers into tapdance bindings, if possible
 - I'd like to have 'launcher keys'
 - I've avoided putting things on WASD because games.  But even on PC
   I'm mostly a controller peon.  I only play FPSes with K&M, and only
   PC-first FPSes like Doom and not like Halo.  I don't play that many
   FPSes either.  But I might make a "gaming layer" that just sends key
   bindings and is otherwise a stock Ergodox layout.