Moved func layer to A/; from R/U
A little less stretching.

Originally had it on G/H but that interferes with holding left in Vim.
I will strive to avoid binding layers to keys I hold down a lot.  This
would be hjkl in Vim-ish contexts, and space, backspace, tab because I
have a bad habit of expecting these to be held down to align code.