Swap from one-shots to vanilla mods
I'm getting a lot of stuck mods, and I am not so good at remembering
what keys I've pressed, so I end up repeatedly pressing mods until
everything comes unglued.

To be honest, I'd be happy if the OSM didn't have multi-tap to hold a
mod --- I will tap once for the one-shot, and I'll hold the key down,
but I almost never multi-tap to hold down.

There's a bug in QMK, issue 3963, that tracks a related issue.

A common recommendation is to integrate Callum Oakley's custom oneshot
implementation, see [1], and [2] for an example of its implementation.
I hear it has quirks of its own though.  I may try this in a free
moment (not any time soon).

[1] https://github.com/callum-oakley/qmk_firmware/blob/master/users/callum/oneshot.h
[2] https://github.com/iandunn/dotfiles/blob/6f87a52fba5340437dca16b7f1fd7d6cabf7eaf6/ergodox-ez/ordinaryish/keymap.c